fix some dependency issues

+++ workforce-service/src/main/java/com/example/domain/model/aggregates/Job.java
@@ -16,7 +16,7 @@ public class Job {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "job_id")
-    private Integer jobId;
+    private Long jobId;
 
     @Column(name = "customer")
     private String customer;
@@ -55,10 +55,10 @@ public class Job {
     // 构造方法、getter、setter
     public Job() {}
 
-    public Integer getJobId() {
+    public Long getJobId() {
         return jobId; }
 
-    public void setJobId(Integer jobId) {
+    public void setJobId(Long jobId) {
         this.jobId = jobId; }
 
     public String getCustomer() {
