+++ workforce-service/src/main/java/com/example/infrastructure/repository/ShiftPlanRepository.java
@@ -7,7 +7,7 @@ import java.util.Date;
 import java.util.List;
 
 public interface ShiftPlanRepository extends JpaRepository<ShiftSchedule,Long> {
-    List<ShiftSchedule> findByEmployeeId(Integer employeeId);
+    List<ShiftSchedule> findByEmployeeId(Long employeeId);
 
     List<ShiftSchedule> findByShiftDate(Date shiftDate);
 }
